Rewrite of notifications
Added capability for e-mail notification of door closing as well as the IFTTT notification.  Used to only get IFTTT notificationon door closures.  You can also opt out of the e-mail or IFTTT notifications by not populating the API for IFTTT or not populating the "from" e-mail address for the e-mail notification.